Refactored subscriber callbacks to allow for updating of the _agent_states dictionary with the agent names
Integrated two callback functions: one for signal and one for car.
Also integrated partial functions for each of these so that the
agent_name can be passed to them so they can be properly stored
in the dictionary.

car_config.yaml
* added newline at the end
car_state.py
* updated subscriber callback to use partial functions to allow
for proper dictionary updating with agent names
* updated for loop and replaced return statement with continue
statement