Reverted to landscape flow chart for final implementation
Reverted to landscape flow chart. This is better for readability, as the
GitHub markdown shows a fullscreen icon in the top right of the
chart. My attempt to make the chart more vertical was successful, however,
it was impossible to view the whole thing on screen at once without a
portrait monitor, which most don't have. This is more accessible.

README.md
* Reverted to previous flow chart design with landscape orientation
* removed image reference in README